requirements.txt
asgiref==3.8.1
confluent-kafka==2.4.0
Django==5.0.6
kafka==1.3.5
kafka-python==2.0.2
mysql-connector==2.2.9
mysql-connector-python==8.4.0
sqlparse==0.5.0
